impl From<Scratch> for PR
