+++ crates/pg_completions/src/context.rs
@@ -1,0 +1,9 @@
+use crate::CompletionParams;
+
+pub struct CompletionContext {}
+
+impl CompletionContext {
+    pub fn new(params: &CompletionParams) -> Self {
+        todo!()
+    }
+}
