+++ crates/pg_completions/src/relevance.rs
@@ -1,0 +1,1 @@
+struct CompletionItemScore {}
